lkdtm: add tests for atomic over-/underflow
This adds additional tests for the remaining atomic functions.
Since the bulk of the logic is identical, the
functions are generated with macros.
Also adds debugfs entries for the manipulated atomic values,
so that the saturation values can be validated after a test is executed.

Based on work by Hans Liljestrand and AKASHI Takahiro.

Signed-off-by: Kees Cook <keescook@chromium.org>